requirements.txt
transformers
datasets
torch
accelerate
bitsandbytes
peft
faiss-cpu
sentence-transformers
rouge_score
nltk
sacrebleu
pyyaml
tqdm
evaluate
matplotlib
seaborn
tabulate
scikit-learn
tensorboard
psutil
gpustat
bert-score
meteor-score
streamlit
rank-bm25
optuna
plotly
pandas
